+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-07-01 12:53:46
-// Commit: test(perf): add unit tests for perf business logic
+// Updated at: 2026-07-02 15:59:15
+// Commit: fix(perf): prevent double tap duplicate action in perf
 
 class perfComponentState {
-  final String commitRef = '2447';
-  final String lastMessage = 'test(perf): add unit tests for perf business logic';
-  final DateTime timestamp = DateTime.parse('2026-07-01T12:53:46+05:00');
+  final String commitRef = '2478';
+  final String lastMessage = 'fix(perf): prevent double tap duplicate action in perf';
+  final DateTime timestamp = DateTime.parse('2026-07-02T15:59:15+05:00');
 }